% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargaMaster.R
\name{descargaMaster}
\alias{descargaMaster}
\title{Es una funcion interna que descarga un archivo en el directorio de trabajo
Es llamada por la funcion \link[sec13f.es]{indice13f}}
\usage{
descargaMaster(link, nombreArchivo, metodoD)
}
\arguments{
\item{link}{direccion donde se encuentra el archivo a descargar}

\item{nombreArchivo}{que se le da al archivo una vez descargado}

\item{metodoD}{metodo de descarga}
}
\value{
descarga el archivo master y devuelve verdadero o falso en funcion de si ha podido descargarlo o no
}
\description{
Es una funcion interna que descarga un archivo en el directorio de trabajo
Es llamada por la funcion \link[sec13f.es]{indice13f}
}
\examples{
\dontrun{

descargaMaster(link, nombreArchivo, metodoD)

}

}
