% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice13f.R
\name{indice13f}
\alias{indice13f}
\title{Carga en la base de datos el indice de formularios}
\usage{
indice13f(periodo = lubridate::year(Sys.Date()),
  nombreBD = paste0(format(Sys.Date(), "\%Y\%m\%d"), "sec13f"),
  mongoURL = "mongodb://localhost:27017")
}
\arguments{
\item{periodo}{el ano en numero entero o vector de n?meros enteros para los que se quiere
descargar el archivo 'master'.}

\item{nombreBD}{nombre de la base de datos de mongo en la que se va a guardar la coleccion que contiene el indice}

\item{mongoURL}{url de la base de datos de mongo, Por defecto, "mongodb://localhost:27017")}
}
\value{
devuelve un dataframe con los datos del indice master
}
\description{
\code{indice13f} descarga la informacion del archivo 'master' trimestral de la SEC para
 los anos solicitados, los une y carga la informacion en la base de datos de mongoDB.
 Si la coleccion esta vacia carga todos los registros del indice y si no esta vacia,
 carga solo los que aun no estan en ese indice en mongoDB.
}
\details{
La funcion indice13f toma como input del usuario el ano o los anos, y descarga
el archivo trimestral 'master' de la SEC que contiene un listado de los formularios enviados.
La direccion de la que lo descarga es : https://www.sec.gov/Archives/edgar/full-index/.
Une los listados de todos los trimestres en uno y lo carga en una coleccion de mongoDB.
Tambien devuelve un dataframe con el contenido que ha cargado en mongo.

El dataframe contiene las siguientes variables:
'accessionNumber' 'cik' 'companyName' 'formType' 'dateFiled' 'edgarLink'

Utiliza la funcion: descargaMaster.R de este paquete
}
\examples{
\dontrun{
master <- indice13f(periodo = 2014:2019,
                    nombreBD = "sec13f",
                    mongoURL = "mongodb://localhost:27017"
                    )

## Descarga un listado de los formularios enviados entre 2013 y 2018.
}

}
