% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiaMaster.R
\name{limpiaMaster}
\alias{limpiaMaster}
\title{Descarga y devuelve el contenido de los formularios de un archivo indice o solo los descarga, a eleccion del usuario}
\usage{
limpiaMaster(coleccion, nombreBD, mongoURL)
}
\arguments{
\item{coleccion}{nombre de la coleccion de mongo que contiene el indice de los formularios a recorrer}

\item{nombreBD}{nombre de la base de datos de mongo que contiene el indice de los formularios a recorrer}

\item{mongoURL}{parametro de la url necesaria para conectar con mongo}
}
\value{
dataframe con los resultados
}
\description{
\link[sec13f.es]{limpiaMaster} Elimina del indice los formularios ya cargados en la base de datos de mongo
Es llamada por la funcion \link[sec13f.es]{recorreFormularios} para evitar duplicar el trabajo y duplicar los registros de formularios.
}
\examples{
\dontrun{
limpiaMaster(coleccion = paste0(format(Sys.Date(),'\%Y\%m\%d'), 'indice'),
             nombreBD = paste0(format(Sys.Date(),'\%Y\%m\%d'), 'sec13f'),
             mongoURL = 'mongodb://localhost:27017')
}

}
