% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorreFormularios.R
\name{recorreFormularios}
\alias{recorreFormularios}
\title{Descarga y devuelve el contenido de los formularios de un archivo indice o solo los descarga, a eleccion del usuario}
\usage{
recorreFormularios(coleccion = paste0(format(Sys.Date(),'%Y%m%d'), 'indice'),
                          db = paste0(format(Sys.Date(),'%Y%m%d'), 'sec13f'),
                          mongoURL = 'mongodb://localhost:27017')
}
\arguments{
\item{coleccion}{nombre de la coleccion de mongo que contiene el indice de los formularios a recorrer}

\item{bd}{nombre de la base de datos de mongo que contiene el indice de los formularios a recorrer}

\item{mongoURL}{parametro de la url necesaria para concectar con mongo}
}
\value{
devuelve el dataframe con el listado de los formularios registrados en ese o esos periodos.
  lo descarga de \url{https://www.sec.gov/Archives/edgar/full-index/}
  Tambien devuelve el estado de la descarga y el numero de formularios e inversores recibidos en ese/esos
  periodos.
}
\description{
\link[sec13f.es]{recorreFormularios} Recorre los formularios de la SEC contenidos en la coleccion que se pasa como input
y almacena la informacion en dos colecciones de documentos de mongo. Una para la tabla de inversiones y otra para el
resto de informacion del formulario 13f. Utiliza \link[sec13f.es]{extraeDatos} para extraer la informacion de cada
formulario.
}
\examples{
\dontrun{
recorreFormularios(coleccion = paste0(format(Sys.Date(),'\%Y\%m\%d'), 'indice'),
                   bd = paste0(format(Sys.Date(),'\%Y\%m\%d'), 'sec13f'),
                   mongoURL = 'mongodb://localhost:27017')
}

}
