% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec13f.es.R
\docType{package}
\name{sec13f.es}
\alias{sec13f.es}
\alias{sec13f.es-package}
\title{sec13.es: Crea una base de datos en mongoDB con los formularios 13f de la SEC}
\description{
El paquete sirve para construir una base de datos en mongoDB a partir de
la informacion online disponible en EDGAR de los formularios 13f de la SEC.
Los formularios 13f contienen informacion de las carteras de inversores en EEUU con
mas de 100 millones de dolares en acciones.
}
\section{funciones en el paquete sec13f}{

indice13f
descargaMaster
extraeDatos
limpiaMaster
recorreFormularios
}

